+++ src/main/java/ar/com/pa/utils/MapperUtils.java
@@ -4,10 +4,12 @@ import java.time.LocalDate;
 
 public interface MapperUtils {
 
-
 	public LocalDate toDate(String date);
+
 	public LocalDate convertToDateYYYY_MM_DD(String year, String month);
-	public int stringToNum(String s);
+
+	public double stringToNum(String s);
+
 	public String addFormat(String old, String next);
-	
+
 }
